update researcher review page as component
